Edit NTASKS logic scripts and provide default values
Assume that a value of 0 for NTASKS means that the "default" value
should be used. Otherwise pass on the value of NTASKS to the
`create_newcase` command.

Currently, mpasa meshes don't have a default value provided by
cam/cime_config/config_pes.xml. This commit also addresses that by
providing our own "default" in the case statments. The default values
are assumed to be:

 - 120km uses 1 node
 - 60km uses 4 nodes (half grid distance implies 4x problem size)
 - 30km uses 16 nodes (quarter grid implies 16x problem size)

Note that the FullyCoupled compset quadruples these numbers