Add script to merge CESM tag into EarthWorks Model
Given a tag of CESM and (optionally if not the default) a path to a
clone of the EarthWorks Model, this script attempts to update all
externals. If the external has been forked to EarthWorksOrg, a merge
occurs. Otherwise if the external is from the same source as in CESM,
the tag is updated in Externals.cfg.

This script creates branches called 'update/{cesm_tag}/{extname}' as
well as new tags (that are incremented from the previous). Currently the
user will have to push these.